Time: 124 ms (27.73%), Space: 133.7 MB (69.55%) - LeetHub

+++ 4036-lexicographically-largest-string-after-pair-transformations/4036-lexicographically-largest-string-after-pair-transformations.java
@@ -2,7 +2,7 @@ class Solution {
     public String[] largestString(int[] nums) {
         int n = nums.length;
         String[] ans = new String[n];
-        long[] arr = new long[26];
+        int[] arr = new int[26];
         arr[0] = 1;
 
         for(int i = 1; i < 26; i++){
@@ -17,7 +17,7 @@ class Solution {
                 for(int j = 25; j >= 0; j--){
                     if(arr[j] <= num){
                         sb.append((char) (j + 'a'));
-                        num = num - (int) arr[j];
+                        num = num - arr[j];
                         break;
                     }
                 }
